% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBD.reproduction.R
\name{bitwise_reproduction}
\alias{bitwise_reproduction}
\title{For parental sampling for Evolutionary Behaviour Dynamics}
\usage{
bitwise_reproduction(fathers, mothers)
}
\description{
Stuff1212
}
\references{
McDowell, J. J. (2004). A computational model of selection by consequences. Journal of the Experimental Analysis of Behavior, 81(3), 297–317. \url{https://doi.org/10.1901/jeab.2004.81-297}
}
