% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAB.EBD.R
\name{CAB.EBD}
\alias{CAB.EBD}
\alias{EBD}
\title{McDowell's Evolutionary Behaviour Dynamics}
\description{
This is a stock implementation of McDowell's Evolutionary Behaviour Dynamics (EBD; McDowell, 2004). See references for a list of papers on EBD.
}
\references{
Kulubekova, S., & McDowell, J. J. (2008). A computational model of selection by consequences: Log survivor plots. Behavioural Processes, 78(2), 291–296. \url{https://doi.org/10.1016/j.beproc.2007.12.005}
Kulubekova, S., & McDowell, J. J. (2013). Computational model of selection by consequences: Patterns of preference change on concurrent schedules. Journal of the Experimental Analysis of Behavior, 100(2), 147–164. \url{https://doi.org/10.1002/jeab.40}
McDowell, J. J. (2004). A computational model of selection by consequences. Journal of the Experimental Analysis of Behavior, 81(3), 297–317. \url{https://doi.org/10.1901/jeab.2004.81-297}
McDowell, J. J. (2010). Behavioral and neural Darwinism: Selectionist function and mechanism in adaptive behavior dynamics. Behavioural Processes, 84(1), 358–365. \url{https://doi.org/10.1016/j.beproc.2009.11.011}
McDowell, J. J. (2013). A quantitative evolutionary theory of adaptive behavior dynamics. Psychological Review, 120(4), 731–750. \url{https://doi.org/10.1037/a0034244}
McDowell, J. J., & Calvin, N. T. (2015). Against matching theory: Predictions of an evolutionary theory of behavior dynamics. Behavioural Processes, 114, 14–25. \url{https://doi.org/10.1016/j.beproc.2015.02.007}
McDowell, J. J., & Caron, M. L. (2007). Undermatching is an emergent property of selection by consequences. Behavioural Processes, 75(2), 97–106. \url{https://doi.org/10.1016/j.beproc.2007.02.017}
McDowell, J. J., Caron, M. L., Kulubekova, S., & Berg, J. P. (2008). A computational theory of selection by consequences applied to concurrent schedules. Journal of the Experimental Analysis of Behavior, 90(3), 387–403. \url{https://doi.org/10.1901/jeab.2008.90-387}
McDowell, J. J., & Popa, A. (2010). Toward a mechanics of adaptive behavior: Evolutionary dynamics and matching theory statics. Journal of the Experimental Analysis of Behavior, 94(2), 241–260. \url{https://doi.org/10.1901/jeab.2010.94-241}
McDowell, J. J., Soto, P. L., Dallery, J., & Kulubekova, S. (2006). A computational theory of adaptive behavior based on an evolutionary reinforcement mechanism, 175. \url{https://doi.org/10.1145/1143997.1144028}
}
