% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAB.extract.R
\docType{methods}
\name{[,CAB.model,character-method}
\alias{[,CAB.model,character-method}
\title{Extract parts of a \code{CAB.model organism} slot}
\usage{
x[v]
}
\arguments{
\item{\code{x}}{A \code{CAB.model} object.}

\item{\code{v}}{String specifying a variable in the \code{organism} slot of \code{x}. Must be a named argument.}
}
\value{
The object in the \code{organism} slot of the specified \code{CAB.model} object.
}
\description{
An extraction method is implemented for the \code{CAB.model} object for extracting objects from the \code{organism} slot. For assignment to the \code{organism} slot, see \code{\link{o_set}}.
}
\note{
Further indexing of the returend object can be done with additional use of the \code{[} function.
}
\examples{
# Using the good_times_model from class.CAB.models
good_times_model

# Get the value of happiness_level
good_times_model[ v = "happiness_level" ]
# Second level indexing. Returns the same thing
good_times_model[ v = "happiness_level" ][1]
Out of range:
good_times_model[ v = "happiness_level" ][2]

}
\seealso{
\code{\link{o_set}} for assignment
}

