% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/samplers.R
\name{mcdowell_sampling}
\alias{mcdowell_sampling}
\alias{srswo}
\alias{srs}
\alias{CAB_samplers}
\alias{CAB_srswo}
\title{For generating random samples}
\usage{
mcdowell_sampling(fitness, fitness_weights)

srswo(choose, repeats, sample_size)

srs(choose_max, repeats, sample_size)
}
\description{
Documentation for samplers.
}
\section{\code{srswo}}{
{
    Simple random sampling without replacemant and uniform selection probabilities
    \subsection{Usage}{
        \code{srswo( choose, repeats, sample_size )}
    }
    \subsection{Arguments}{
        \describe{
            \item{\code{choose}}{Numeric vector of elements to choose from.}
            \item{\code{repeats}}{Integer. How many repeats for the sampling.}
            \item{\code{sample_size}}{Integer. How many elements chosen per sample.}
        }
    }
    \subsection{Value}{
        Returns a \code{matrix} with the draws from each sample along the columns.
    }
}
}

\section{\code{mcdowell_sampling}}{
{
    Sampling used in the supplementary material of McDowell (2013).
    \subsection{Usage}{
        \code{CAB.mcdowell_sampling( fitness, fitness_weights )}
    }
    \subsection{Arguments}{
        \describe{
            \item{\code{fitness}}{Numeric vector of fitness values.}
            \item{\code{fitness_weights}}{Numeric vector of fitness weights.}
        }
    }
    \subsection{Value}{
        Returns a \code{matrix} with two columns and rows equal to the number of fitnss values. This matrix contains the indices of the values that are sampled. The indexing starts from zero.
    }
}
}

