% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_params.R
\name{change_params}
\alias{change_params}
\title{Change model parameters}
\usage{
change_params(mutables, new_params, model_object)
}
\arguments{
\item{mutables}{A named list. The name of each element should refer to a \code{control} object and the elements should be the names of arguments for the associated \code{elemental} function. These are the parameters that will be changed.}

\item{new_params}{A vector of new parameter values. The order of the vector should correspond to the order of arguments in \code{mutables}.}

\item{model_object}{A \code{model} object or any of its child classes.}
}
\description{
The \code{change_params()} function is for changing the parameters in a \code{model} object that has already been constructed. Note that this function is implemented only for \code{model} objects that have one \code{elemental} function for each \code{control} slot. In the future, this may be changed.
}
\examples{
# Suppose a model, "model_obj", has a "DOR_control" slot with one parameter "max" that is initially set to 100.
# To change the value to 200:
mutables = list( DOR_control = "max" )
new_params = 200
change_params( mutables, new_params, model_obj )

}
\seealso{
\code{\link{class.model}}
}

