% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.event.R
\docType{methods}
\name{check.event}
\alias{check.event}
\alias{check.event,input,character,character-method}
\title{Check whether an event has occurred or not}
\usage{
check.event(input_object, event_var, setup_var)

check.event( input_monitor, event_var, setup_var )
}
\arguments{
\item{event_var}{A character string speicfying the event variable in the \code{sim_input} object.}

\item{setup_var}{A character string specifying the variable that sets up the event in the \code{sim_input} object.}

\item{input_monitor}{A \code{sim_input} object.}
}
\value{
Returns \code{TRUE} if the event has occurred.
}
\description{
Check whether or not an event that has been arranged has occurred or not. This function can be used for any function that specifies something in the future.
}
\examples{
# From the ?do.sim_fx example
set.seed(1)
access.input_monitor( "rft_setup", 1, do.sim( event, input_monitor ) )
check.event( input_monitor = input_monitor, event_var = "rft_time", setup_var = "rft_setup" )

}
\seealso{
\code{\link{input_monitor}} For how the events are recorded and set up.
}

