% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{class}
\name{dataset-class}
\alias{class.dataset}
\alias{dataset-class}
\alias{make.custom_dataset_object}
\alias{set.custom_dataset}
\title{The \code{dataset} class.}
\usage{
set.custom_dataset(class_name, slotnames)

make.custom_dataset_object(dataset_class, ...)
}
\description{
In the CAB package, experimental data that is imported for analysis is stored in an \code{analysis_object} or one of its child classes (\code{\link{class.analysis_object}}. For the analysis of experimental data, each session should be stored in separate \code{analysis_object}s and these \code{analysis_object}s are in turn contained in a \code{dataset} object.
}
\section{Slots}{

\describe{
\item{\code{meta_data}}{A \code{data.frame} that contains the information about each element (i.e. experimental session) in the \code{analysis_objects} slot.}

\item{\code{analysis_object}}{A list, where each element is an \code{analysis_object}.}
}}
\section{{Custom \code{dataset} class}}{
{
If the user wishes to construct their own \code{dataset} class, the \code{set.custom_dataset()} function can be used.

To create an object of the newly set \code{dataset} child class, the \code{make.custom_dataset_object()} function can be used.
#'     \subsection{Usage}{
        \code{set.custom_dataset( class_name, slot_names )}
    }
    \subsection{Mandatory slots}{
        \describe{
            \item{\code{analysis_objects}}{Inherited from \code{analysis_object}.}
            \item{\code{meta_data}}{Inherited from \code{meta_data}.}
        }
    }
    \subsection{Arguments}{
        \describe{
            \item{\code{class_name}}{A character string giving the name of the custom \code{dataset} child class.}
            \item{\code{slot_names}}{Either: 1) an unnamed list of slot names or 2) a named list, with the names specifying the slot names and the elements specifying the class of object contained in the slot. If 1), all slots will contain \code{list}.}
        }
    }
    \subsection{Usage}{
        \code{make.custom_dataset_object( dataset_class, ... )}
    }
    \subsection{Arguments}{
        \describe{
            \item{\code{dataset_class}}{A character string specifying the name of the custom \code{dataset} class.}
            \item{\code{...}}{A named set of arguments that fills up the object. The arguments should be named as the slot name to which the argument pertains to.}
        }
    }
}
}
\seealso{
\code{\link{analysis_object}} For analogous examples for the usage of \code{set.custom_dataset} and \code{make.custom_dataset_object}.
}

