% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dor.R
\docType{class}
\name{DOR-class}
\alias{DOR}
\alias{DOR-class}
\alias{DOR.show}
\alias{class.DOR}
\alias{remove.DOR.show}
\title{Delay of Reinforcement Gradient (DOR) class}
\format{An object of class \code{character} of length 1.}
\usage{
DOR.show

remove.DOR.show()
}
\description{
Delay of Reinforcement Gradients are represented as elements in the \code{DOR} class. The use of DORs can be found in Catania (2005) and Berg & McDowell (2011). The \code{DOR} class is used to set up the DOR function, that is, its parameters are not set here. Instead, the \code{DOR} will go into a \code{DOR_control} object that contains both the \code{DOR} as well as its associted parameters.
}
\details{
In the CAB package, objects from classes like \code{DOR} are called 'elemental' objects That is, they comprise a particular element of the simulation without any specified parameter values. 'Control' objects are elemental objects with an associated list of parameter values.

The way to put a DOR in your model is as follows:
\enumerate{
    \item Code up the function that you want to use for the DOR or use an existing function.
    \item Construct a \code{DOR} object from the function.
    \item Make a list specifying the values of the parameters for the DOR function.
    \item Make a \code{DOR_control} object, which contains the \code{DOR} object and the parameters. See \code{\link{make.control}}.
}

The \code{show} method for the \code{DOR} class can be removed with the function \code{remove.DOR.show()} and reinstated with \code{DOR.show}.
}
\section{Slots}{

\describe{
\item{\code{DOR}}{This will contain a DOR function. See examples.}

\item{\code{name}}{The name of the \code{DOR} object. Inherited from \code{elemental}.}

\item{\code{type}}{This will be \code{elemental} because the \code{DOR} is an \code{elemental} object. Inherited from \code{elemental}.}
}}
\section{Built-in DOR functions}{
{
Here is a list of the built-in DOR functions. The function definitions can be seen by calling the name of the function without brackets. For each of these functions, \eqn{t} is the time difference between a response and the latest reinforcement. Each function takes three parameters, \eqn{t}, \eqn{max} as the maximum reinforcement effect of an individual behaviour, and \eqn{scale} as a scaling parameter of the DOR.
    \describe{
        \item{\code{linear_DOR_fx}}{A DOR where the reinforcement effect is a linear function of time since latest reinforcement. \deqn{ f(t) = max - 1/scale * t}
        Note that the reinforcemet effect is bounded by zero.}
        \item{\code{exponential_DOR_fx}}{A DOR where the reinforcement effect is an exponential function of the time since latest reinforcement. \deqn{ f(t) = max * exp( -t/scale ) }}
        \item{\code{reciprocal_DOR_fx}}{A DOR where the reinforcement effect is a reciprocal function of the time since latest reinforcement. \deqn{ f(t) = max / ( t+1 )^( 1/scale ) }}
        \item{\code{hyperbolic_DOR_fx}}{A DOR where the reinforcement effect is a hyperbolic function of the time since latest reinforcement. \deqn{ f(t) = 2 * max / ( ( t + 1 )^( 1/scale ) + 1 )}}
    }
With respect to the actual implementation of these DOR functions, each of the built-in DOR functions have five arguments.
\describe{
    \item{\code{max}}{The maximum increment from an individual response.}
    \item{\code{scale}}{A scaling factor on the DOR. The scaling is independent of \code{max}.}
    \item{\code{IRI_resp_times}}{A vector containing response times counting backward from the most recent reinforcement time.}
    \item{\code{reserve}}{A numeric specifying the value of the reserve.}
    \item{Each of these arguments can instead be a string specifying the variable that should be taken as the argument from a \code{input} object.}{}
}
The DOR function will return the value of the reserve after the increment.
}
}

\section{Make a \code{DOR} object from a DOR function}{
{
Use the \code{make.DOR} function to make a \code{DOR} object.
    \subsection{Usage}{
        \code{make.DOR(DOR_fx, name)}
    }
    \subsection{Arguments}{
        \describe{
            \item{\code{DOR_fx}}{A function that represents the DOR of interest}
            \item{\code{name}}{A character vector specifying the name of your \code{DOR} object}
        }
    }
    \subsection{Value}{
        Returns a \code{DOR} object.
        }
}
}
\examples{



}
\keyword{datasets}

