% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_input.R
\docType{class}
\name{sim_input-class}
\alias{class.sim_input}
\alias{make.sim_input}
\alias{sim_input-class}
\alias{sim_input.show}
\title{Simulation input class}
\format{The \code{show} method prints the \code{sim_input} object as a \code{data frame}, which is what it would print out as without the \code{show} method.}
\usage{
make.sim_input(...)

sim_input.show
}
\arguments{
\item{...}{The names of the desired \code{input} objects, separated by commas.}
}
\description{
Objects of the \code{sim_input} class store objects of the \code{input} class (see \code{\link{class.input}}. The use of the \code{sim_input} class is to put all of the necessary \code{input} objects together for a simulation. During the simulation, the \code{sim_input} object will contain the values of the inputs at any given time.
}
\details{
\code{make.sim_input} for making a \code{sim_input} object.

There is a \code{show} method for the \code{sim_input} class. This method can be removed with the \code{remove.sim_inout.show} function and reinstated with the \code{sim_input.show}. This show method does not do anything different to if the object were just printed normally.

In a simulation, \code{sim_input} is used to monitor the value of the inputs, which serve as arguments for the functions that are necessary to operate a given model. See the section below. See \code{\link{input_monitor}}.
}
\section{Slots}{

\describe{
\item{\code{input}}{This is a list of the input objects that are required for a simulation.}
}}
\section{Make a simulation input object}{
{
To create an object of \code{sim_input}, use the function \code{make.sim_input}.
}
}
\examples{
# Make some input objects
test_input1 = make.custom_input( "test_input1" )
test_input2 = make.custom_input( "test_input2" )
# Make a sim_input object
make.sim_input( test_input1, test_input2 )

}
\seealso{
\code{\link{make.custom_input}} to make \code{input} objects.

\code{\link{input_monitor}} for when \code{sim_input} is used in a simulation.
}
\keyword{datasets}

