% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_catalogue.R
\name{class_catalogue}
\alias{class_catalogue}
\title{View the objects in a class}
\usage{
class_catalogue(class_name, custom_env = NULL)
}
\arguments{
\item{class_name}{A string that gives the name of the class that you want to view the objects of.}

\item{environment}{A character vector that gives the environments that you want to search for objects in. The default value is \code{NULL} and by default, the \code{class_catalogue} searches through the global environment and also the \code{CAB:package} environment.}
}
\value{
Returns a character vector containing the names of all the instances of the class of interest.
}
\description{
This function \code{class_catalogue} is for getting the names of objects from a specified class. It is most useful for when you want to view the objects that you have defined. For example, the \code{\link{make.custom_input}} function should be used to define inputs with unique names. \code{class_catalogue} can be used to check if your custom input names are unique.
}
\examples{
# Make a test class
test_input = make.custom_input( name = "test_input" )
# Check that it exists
class_catalogue( "input" )

}

