% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_IxyI.R
\name{clean_short_IxyI}
\alias{clean_short_IxyI}
\title{Clean your data}
\usage{
clean_short_IxyI(data, x_event, y_event, gap)
}
\arguments{
\item{data}{An event record. At the moment, methods are defined only for "formal_event_record".}

\item{x_event}{A character string specifying the first event type.}

\item{y_event}{A character string specifying the second eevnt type. Can be missing.}

\item{gap}{A numeric giving the gap between \code{x_event} and \code{y_event}}
}
\description{
A method for cleaning short inter-x-y-intervals.
}
