% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.I_xy_I.R
\docType{methods}
\name{compute.I_xx_I.analysis_object}
\alias{compute.I_xx_I.analysis_object}
\alias{compute.I_xx_I.sim_analysis_object}
\alias{compute.IxyI,dataset-method}
\alias{inter-response_time}
\title{Calulating the time elapsed between events in a simulation}
\usage{
compute.I_xx_I.analysis_object(data, x_event, break_event, x_offset = 0)

compute.I_xx_I.sim_analysis_object(data, x_event, break_event, x_offset = 0)

\S4method{compute.IxyI}{dataset}(data, x_event, x_offset = 0,
  y_event = "NONE", break_event = "ALL")
}
\arguments{
\item{data}{A \code{analysis_object} or \code{simulation_analysis_object}.}

\item{x_event}{A character string specifying the "x" in "IxyI".}

\item{break_event}{A character vector specifying the break events over which the "IxyI" or "IxxI" cannot be computed over. Defaults to \code{"ALL"} for which only the "IxyI", where "y" must immediate follow "x" is computed. Also takes \code{"NONE"} for which the "IxyI" is calculated for all "y" that follow "x" regardless of the intervening events.}

\item{x_offset}{A numeric specifying the duration of the "x". Defaults to 0.}

\item{y_event}{A character specifying the "y" in "IxyI". Defaults to \code{NULL}, which means the function computes the "IxxI".}
}
\description{
Given some data from either an experiment or a simulation, the time that elapsed between two kinds of specified events is computed for all instances of the pairing of the two events. The generic description of this elapsed time is the "IxyI" time, read as the "inter-x-y-interval". When x=y, and x = a time at which responses occurred, then the IxyI is an inter-response time (IRT). For another example, when x != y, x = a time at which a reinforcement delivery occurred and y = a time at which a response occured, then the IxyI is a post-reinforcement pause. Functionality is included for skipping over specific events (such as Computing IRTs, while excluding the IRTs with an intervening reinforcement delivery). Further functionality is included for adding an offset to the \code{x_event}, for example, if computing the post-reinforcement pause, it is desirable to subtract out the reinforcement delivery.
}
\details{
With respect to efficiency, the \code{compute.IxyI} method for \code{simulation_analysis_object} is anywhere between twice to five times as fast as the method for \code{analysis_object}. However, to achieve a noticible difference in actual time, the functions would need to be run on the order of 10000 times.

In cases where there is only once instance of a particular event, the value returned is a single \code{Inf}. An example is for the calculation of the inter-response times when only once response occurs during a simulation. Another example is is for the calculation of post-reinforcement pauses, but no responses follow a reinforcement delivery.

When the \code{compute.IxyI} is called for objects of class \code{dataset.R}, a list of the requested statistic is computed.
}
\examples{
#Suppose "d" is your data in an "analysis_object", "resp_time" is the response time, "rft_time" is the reinforcement time
#Compute the inter-response times, including those with intervening reinforcement deliveries
irts_inc_rft = compute.IxyI( d, x_event = "resp_time" )
#Compute the IRTs excluding those with intervening rft deliveries
irts = compute.IxyI( d, x_event = "resp_time", break_event = "rft_time" )

#Compute post-rft pauses including rft duration
prp = compute.IxyI( d, x_event = "rft_time", y_event = "resp_time" )
#Compute prp excluding rft duration, if rft duration is 3
prp_less_rft = compute.IxyI( d, x_event = "rft_time", y_event = "resp_time", x_offset = -3 )

}
\seealso{
\code{\link{class.analysis_object}} For constructing arguments for \code{data} parameter.
}

