% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.session_rates.R
\docType{methods}
\name{compute.session_rates}
\alias{compute.session_rates}
\alias{compute.session_rates,formal_event_record,list,character,numeric-method}
\alias{compute.session_rates,formal_event_record,list,missing,numeric-method}
\alias{compute.session_rates,formal_event_record,list,missing,missing-method}
\alias{compute.session_rates,formal_event_record,list,character,missing-method}
\alias{compute.session_rates,ragged_event_record,list,character,numeric-method}
\alias{compute.session_rates,ragged_event_record,list,missing,numeric-method}
\alias{compute.session_rates,ragged_event_record,list,character,missing-method}
\alias{compute.session_rates,ragged_event_record,list,missing,missing-method}
\title{Compute session rates}
\usage{
compute.session_rates(data, event_offsets, dims = NULL,
  session_duration = NULL)


  \S4method{compute.session_rates}{formal_event_record,list,character,numeric}(data,
  event_offsets, dims = NULL, session_duration = NULL)


  \S4method{compute.session_rates}{formal_event_record,list,missing,numeric}(data,
  event_offsets, session_duration)


  \S4method{compute.session_rates}{formal_event_record,list,missing,missing}(data,
  event_offsets)


  \S4method{compute.session_rates}{formal_event_record,list,character,missing}(data,
  event_offsets, dims = NULL, session_duration = NULL)


  \S4method{compute.session_rates}{ragged_event_record,list,character,numeric}(data,
  event_offsets, dims = NULL, session_duration = NULL)


  \S4method{compute.session_rates}{ragged_event_record,list,missing,numeric}(data,
  event_offsets, session_duration)


  \S4method{compute.session_rates}{ragged_event_record,list,character,missing}(data,
  event_offsets, dims)


  \S4method{compute.session_rates}{ragged_event_record,list,missing,missing}(data,
  event_offsets)
}
\arguments{
\item{data}{Either an object of class \code{simulation_analysis_object}, \code{analysis_object}, or \code{dataset}.}

\item{dims}{A character vector of variable names to calculate the rates of. Defaults to \code{NULL}. If \code{NULL}, the rates of all the events in the \code{data} argument are computed.}

\item{session_duration}{A numeric value specifying the duration of the session over which the rates are calculated. Defaults to \code{NULL}. If \code{NULL}, the latest time associated with an event is taken as the session duration. See details.}

\item{rft_duration}{A list with each element being a numeric specifying the reinforcemnt duration and the name of each element corresponding to the reinforcement event.}
}
\description{
\code{compute.session_rates} is a method for calculating the within-session rate of some variable. Currently, methods are defined for the classes \code{analysis_object} and its child class \code{simulation_analysis_object}.
}
\details{
\subsection{\code{session_duration = NULL}}{
When \code{data} is of class \code{simulation_analysis_object}, the session duration is taken as the maximum recorded time for the list of variables for which rates are calculated. Note that the variables associated with reinforcement delivery in \code{rft_duration} are automatically included in determining the session duration.

When the \code{data} is of class \code{analysis_object}, the session duration is taken as the time associated with the last entry in the \code{event_time} data frame.

When the \code{data}is of class \code{dataset}, the \code{compute.session_rates} method for the \code{analysis_object} is just looped over all the \code{analysis_object}s contained in the \code{dataset} object.
}

\subsection{The computation of the rate}{
When the rate is calculated, regardless of the method, the number of times that a particular event occurs is divided by the session time. The session time is corrected by the time that is taken up by reinforcement deliveries. The session time taken up by reinforcement is calculated as \eqn{rft_time * rft_number}. A further correction is done when the last event in the simulation or experiment is a reinforcement delivery, where the session time taken by reinforcement is calculated as \eqn{rft_time * (rft_number-1)}.
}

\subsection{Computing rates for objects of \code{dataset}}{
To calculate the rates for the multiple sessions that are stored in a \code{dataset} object, we essentially call \code{lapply} over all of the \code{analysis_objct}s in the \code{dataset}.
}

With respect to efficiency, the \code{compute.session_rates} method for \code{simulation_analysis_object} is about 5 times faster than the method for \code{analysis_object}. For standard data analysis, where time is not a big concern, either method is fine. In fact, the method using \code{analysis_object} may be better because it uses the \code{event_time} dataframe in the parent \code{analysis_object} class, so it is functionally more convenient. When doing simulations, time is a concern. Hence, the method with \code{simulation_analysis_object} should be used. Hopefully, this gives the reader some understanding about the design rationale for this method.
}
\seealso{
\code{\link{class.analysis_object}} For constructing arguments for \code{data} parameter.
}
