% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_event_record.R
\name{convert_event_record}
\alias{convert_event_record}
\title{Convert between \code{ragged_event_record} and \code{formal_event_record}}
\usage{
convert_event_record(event_record, dims = NULL)
}
\arguments{
\item{event_record}{An object of class \code{ragged_event_record} or \code{formal_event_record}.}

\item{dim}{An argument when the \code{event_record} is \code{event_record}. A character vector specifying the variables to retain and the order in which they are retained.

If \code{event_record} is a \code{ragged_event_record}, it will be converted to a \code{format_event_record}, and vice versa.

The ordering of the \code{dim} arguemnt is very important. Often, events may be recorded at the same time but there exists a causal relation between them. For example, response-contingent reinforcement will result in reinforcement and the response being recorded at the same time. To ensure that an \code{formal_event_record} is generated such that responses are placed before their associated reinforcement deliveries, order the \code{dims} argument as such: \code{dims = c("resp_time", "rft_time" )}.}
}
\description{
\code{convert_event_record} is for converting objects of \code{ragged_event_record} to \code{formal_event_record} or vise versa.
}
\seealso{
\code{\link{class.event_record}}
}

