% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_event_record.R
\name{copy_event_record}
\alias{copy_event_record}
\title{Copy a \code{event_record}}
\usage{
copy_event_record( event_record )
}
\arguments{
\item{event_record}{Either a \code{formal_event_record} or \code{ragged_event_record}.}
}
\description{
A \code{event} record is typically implemented as an \code{environment} type class. This means that they have pass-by-reference semantics. As a result, assignment will not produce a copy of the object. Therefore, changing the object from the assigned value will change the original version.
}
