% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAB.COR_helpers.R
\docType{data}
\name{COR_function_names}
\alias{COR.G_E_emission}
\alias{COR.constant_depletion}
\alias{COR.exponential_vi}
\alias{COR.initial_reserve}
\alias{COR.linear_DOR}
\alias{COR.rft_duration}
\alias{COR_function_names}
\alias{DOR}
\alias{depletion}
\alias{initial_reserve}
\alias{rft_duration}
\title{Catania's Operant Reserve functions}
\format{An object of class \code{list} of length 1.}
\usage{
COR_function_names

COR.linear_DOR(DOR_max, DOR_scale, IRI_resp_times, reserve_value)

COR.constant_depletion(depletion_constant, reserve_value)

COR.initial_reserve(initial_reserve_constant)

COR.exponential_vi(inter_rft_interval, time)

COR.G_E_emission(reserve_value, time, min_irt)

COR.rft_duration(time, food_duration)
}
\description{
A list of the functions in the \code{CAB} package's implementation of Catania's Operant Reserve. See \link{CAB.COR} for more details.
}
\details{
Stuff
}
\section{DOR functions}{
{
    A delay-of-reinforcement gradient. Replenishes the reserve.
    \subsection{\code{COR.linear_DOR}}{
        A linear DOR.
    }
    \subsection{Usage}{
        \code{COR.linear_DOR( DOR_max, DOR_scale, IRI_resp_times, reserve_value )}
    }
    \subsection{Arguments}{
        \describe{
            \item{\code{DOR_max}}{A numeric giving the maximum value of the DOR.}
            \item{\code{DOR_scale}}{A numeric that is proportional to the slope of the DOR.}
            \item{\code{IRI_resp_times}}{A numeric vector of response times in the most recent inter-reinforcement interval}
            \item{\code{reserve_value}}{A numeric giving the reserve value.}
        }
    }
    \subsection{Details}{
        The reserve value is defined on [0,1] for our COR implementaiton. Therefore, \code{COR.linear_DOR} will never return a value larger than 1.
    }
    \subsection{Value}{
        Returns the new value of the reserve after applying the DOR.
    }
}
}

\section{Depletion functions}{
{
    A reserve depletion function. Depletes the reserve.
    \subsection{\code{COR.constant_depletion}}{
        A constant depletion function.
    }
    \subsection{Usage}{
        \code{COR.constant_depletion( depletion_constant, reserve_value )}
    }
    \subsection{Arguments}{
        \describe{
            \item{\code{depletion_constant}}{A numeric giving the depletion of the DOR.}
            \item{\code{reserve_value}}{A numeric giving the reserve value.}
        }
    }
    \subsection{Details}{
        The reserve value is defined on [0,1] for our COR implementaiton. Therefore, \code{COR.constant_depletion} will never return a value smaller than 0.
    }
    \subsection{Value}{
        Returns the new value of the reserve after applying the depletion function.
    }
}
}

\section{Inital reserve functions}{
{
    Specify the initial reserve value.
    \subsection{\code{COR.initial_reserve_constant}}{
        A constant initial reserve value.
    }
    \subsection{Usage}{
        \code{COR.initial_reserve( initial_reserve_constant )}
    }
    \subsection{Arguments}{
        \describe{
            \item{\code{initial_reserve_constant}}{A numeric giving the initial reserve value.}
        }
    }
    \subsection{Details}{
        The reserve value is defined on [0,1] for our COR implementaiton. Therefore, \code{COR.initial_reserve_constant} should be in [0,1].
    }
    \subsection{Value}{
        Returns the initial reserve value.
    }
}
}

\section{Reinforcement schedule functions}{
{
    Control the delivery of reinforcement
    \subsection{\code{COR.exponential_vi}}{
        A variable-interval schedule where the inter-reinforcement intervals are arranged by sampling from an Exponential(1/rate) distribution.
    }
    \subsection{Usage}{
        \code{COR.exponential_vi( inter_rft_interval, time )}
    }
    \subsection{Arguments}{
        \describe{
            \item{\code{inter_rft_interval}}{A numeric giving the average inter-reinforcement interval.}
            \item{\code{time}}{A numeric giving the time.}
        }
    }
    \subsection{Value}{
        Samples a new inter-reinforcement interval and adds that to the current \code{time}. Therefore, returns the time when the next reinforcement will be primed.
    }
}
}

\section{Behaviour emission functions}{
{
    Emit responses from the reserve.
    \subsection{\code{COR.G_E_emission}}{
        A function that emits responding based on the transformation from a Geometric distribution to an Exponential distribution.
    }
    \subsection{Usage}{
        \code{COR.G_E_emission( reserve_value, time, min_irt )}
    }
    \subsection{Arguments}{
        \describe{
            \item{\code{reserve_value}}{A numeric giving the reserve value.}
            \item{\code{time}}{A numeric giving the time.}
            \item{\code{min_irt}}{A numeric that is the minimum inter-response time.}
        }
    }
    \subsection{Details}{
        Under standard implementations of COR ( see \link{CAB.COR} ), the reserve gives the probability of responding at a given time point. Therefore, the distribution of how many time intervals to wait until the next response is a Geometric(p) distribution. We transform the Geometric distribution to an Exponential distribution so that responses can be emitted in continuous time. The distribution of each inter-response time is \deqn{ t ~ Exponential( rate = -log(1-reserve) ) + min_IRT } where \eqn{t} is the emitted inter-response time.
    }
    \subsection{Value}{
        Returns the time at which the next response will occur.
    }
}
}

\section{Reinforcement delivery functions}{
{
    Deliver reinforcement.
    \subsection{\code{COR.rft_duration}}{
        A function that accounts for the time associated with a reinforcement delivery.
    }
    \subsection{Usage}{
        \code{COR.rft_duration( time, food_duration )}
    }
    \subsection{Arguments}{
        \describe{
            \item{\code{time}}{A numeric giving the time.}
            \item{\code{food_duration}}{A numeric that is the minimum inter-response time.}
        }
    }
    \subsection{Value}{
        Returns the time after the reinforcement delivery
    }
}
}
\seealso{
\link{CAB.COR}
}
\keyword{datasets}

