% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/EBD.reproduction.R
\name{EBD_BR}
\alias{EBD_BR}
\alias{EBD_RS_BR}
\alias{EBD.reproduction}
\title{Sampling}
\usage{
EBD_BR(fathers, mothers)

EBD_RS_BR(parents)
}
\description{
Stuff1212
}
\references{
McDowell, J. J. (2004). A computational model of selection by consequences. Journal of the Experimental Analysis of Behavior, 81(3), 297–317. \url{https://doi.org/10.1901/jeab.2004.81-297}
}
