% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBD_fitness_def.R
\docType{class}
\name{EBD_fitness_def-class}
\alias{EBD_fitness_def-class}
\alias{class.EBD_fitness_def}
\title{Evolutionary Behaviour Dynamics Fitness Definition class}
\description{
In Evolutionary Behaviour Dynamics (McDowell, 2004), each behaviour has an associated fitness. This determines the probability with which they are selected to be parents. Behaviours in the population are sampled to be parents from a probability distribution (fitness function) with its domain defined by fitness values. The fitnesses definition is implemented as a \code{elemental} object that contains a function that calculates the fitnesses according to the specified fitness definition.
}
\section{Slots}{

\describe{
\item{\code{EBD_fitness_def}}{Contains a function for calculating fitness values.}

\item{\code{name}}{The name of the \code{EBD_fitness_def} object. Inherited from \code{elemental}.}

\item{\code{type}}{This will be \code{elemental} because \code{EBD_fitness_def} is a \code{elemental} object. Inherited from \code{elemental}.}
}}
\section{Built-in fitness definitions}{
{
Here is a list of the built-in fitness definition functions.
    \describe{
        \item{\code{wrapped_specific_individual_fitness}}{Fitness is defined as the minimum absolute difference between each behaviour in the population and a reference behaviour, with the distance wrapped around a circle.
        }
    }
}
}

