% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBD_fitness_fx.R
\docType{class}
\name{EBD_fitness_fx-class}
\alias{EBD_fitness_fx-class}
\alias{class.EBD_fitness_fx}
\title{Evolutionary Behaviour Dynamics Parental Fitness Function class}
\description{
In Evolutionary Behaviour Dynamics (McDowell, 2004), each behaviour has an associated fitness. This determines the probability with which they are selected to be parents. Behaviours in the population are sampled to be parents from a probability distribution (fitness function) with its domain defined by fitness values. The fitness function is implemented as a \code{elemental} object that contains the fitness function.
}
\details{
Do this later
}

