% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBD_sample.R
\name{EBD_sample}
\alias{EBD_sample}
\title{Emit a behaviour from an EBD population}
\usage{
EBD_sample(pop, size = 1, replace = T, prob = NULL)
}
\arguments{
\item{pop}{A \code{EBD_pop} object.}

\item{size}{An \code{integer} specifying the number of behaviours to sample. Defaults to 1.}

\item{replace}{A \code{logical} specifying whether to sample with repalcement. Defaults to \code{TRUE}.}

\item{prob}{A \code{numeric} vector specifying the sampling probabilities for each behaviour. Defaults to \code{NULL} for equal probabilityies.}
}
\value{
Returns a vector of integers sampled from the \code{EBD_pop} genotypes.
}
\description{
In Evolutionary Behaviour Dynamics (McDowell, 2004), a behaviour is emitted from the algorithm by sampling a behaviour from the population with equal probability. The \code{EBD_sample} function is simply a wrapper for the \code{sample} function.
}
\references{
McDowell, J. J. (2004). A computational model of selection by consequences. Journal of the Experimental Analysis of Behavior, 81(3), 297–317. \link{https://doi.org/10.1901/jeab.2004.81-297}
}
\seealso{
\code{\link{sample}}
}

