% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/EBD_utilities.R
\name{EBD_B_premutate}
\alias{EBD_B_premutate}
\alias{bin2int}
\alias{int2bin}
\title{Evolutionary Behaviour Dynamics Utilities}
\usage{
EBD_B_premutate(mutation_rate, pop_size, total_ticks, n_bits)

bin2int(binaries, digits)

int2bin(integers, digits)
}
\description{
This is the documentation for the EBD utility functions.
}
\section{\code{bin2int}}{
{
    Convert a binary string from a binary vector to an integer. Works for matrices where each column is a binary string.
    \subsection{Usage}{
        \code{bin2int( binaries )}
    }
    \subsection{Arguments}{
        \describe{
            \item{\code{binaries}}{A \code{numeric} vector where each element corresponds to a binary digit, or a \code{matrix} where each column corresponds to a binary digit.}
        }
    }
    \subsection{Value}{
        Returns a \code{numeric} vector of the corresponding integers
    }
}
}

\section{\code{int2bin}}{
{
    Converts integers to their corresponding binary representation. Note that the binary string is reversed. For example \code{1} is \code{10} as opposed to \code{01}.
    \subsection{Usage}{
        \code{int2bin( integers, digits )}
    }
    \subsection{Arguments}{
        \describe{
            \item{\code{integers}}{A \code{numeric} vector of integers}
            \item{\code{digits}}{The number of digits for the binary string}
        }
    }
    \subsection{Value}{
        Returns a \code{matrix} where each column is a binary string and each row is the digit on the given string.
    }
}
}

