% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.model_conditions.R
\name{get.model_conditions}
\alias{get.model_conditions}
\title{Get the conditions from a \code{model} oject.}
\usage{
get.model_conditions(model_object, i)
}
\arguments{
\item{model_object}{A model object}

\item{i}{A numeric specifying which condition to get}
}
\description{
A \code{model} object will have a number of conditions that depends on the factorial combination of all the possible parameters. The \code{get.model_conditions} function is used to get the conditions at a particular factorial combination. The main use of this function is for doing factorial simulations.
}
\details{
\code{get.model_conditions} is useful for getting an argument for the function \code{\link{make.sim_functions}}.
}
\examples{
# Following from the example in ?class.model.COR
my_COR
get.model_conditions( my_COR, 1 )
get.model_conditions( my_COR, 2 )

}
\seealso{
\code{\link{class.model.COR}} For the example.

\code{\link{make.sim_functions}} Where \code{get.model_conditions} is a nice way to specify the \code{conditions} argument for \code{make.sim_functions}.
}

