% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_matrix.R
\name{indicator_matrix}
\alias{indicator_matrix}
\title{Construct an indicator matrix}
\usage{
indicator_matrix( x )
}
\arguments{
\item{x}{A \code{numeric} vector}
}
\value{
A matirx with the rows as the length of \code{x} and columns the number of unique \code{x} values.
}
\description{
From a vector \code{x}, construct an indicator matrix
}

