% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_duplicated.R
\name{not_duplicated}
\alias{not_duplicated}
\title{Find not duplicated values}
\usage{
not_duplicated(x)
}
\description{
Calls the \code{duplicated} function in \code{Rcpp} and does a logical inversion. This returns a logical value giving \code{TRUE} for values that are not duplicated and \code{FALSE} otherwise.
}

