% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_elemental.R
\name{set.custom_elemental}
\alias{make.custom_elemental}
\alias{remove.show.custom_elemental}
\alias{set.custom_elemental}
\title{Create a custom elemental class and custom elemental objects}
\usage{
set.custom_elemental(class_name, set_show_method = T)

make.custom_elemental(custom_elemental_fx, custom_elemental_name, class)

remove.show.custom_elemental(object_name)
}
\arguments{
\item{class_name}{A character string specifying the name of the custom elemental class. The name of the slot in the custom elemental object that contains the elemental function will be the same as the \code{class} name.}

\item{custom_elemental_fx}{A function for the elemental object.}

\item{custom_elemental_name}{A name for the elemental object.}

\item{class}{The type of elemental that you want your elemental object to be.}
}
\description{
For the implementation of a custom model, it is necessary to define a custom elemental class as well as objects for that class. When a custom elemental class is defined, a \code{show} method is also defined at the same time by default. To remove the assocaited \code{show} method, use \code{remove.show.custom_element( "class_name" ) }.
}
\note{
If there are no errors, the \code{set.custom_elemental} function will print a message saying that the custom elemental class has been set.
}
\examples{
#Make a DOR class from scratch
set.custom_elemental( class_name = "DOR2" )
#use linear_DOR_fx()
linear_DOR = make.custom_elemental( linear_DOR_fx, "linear_DOR", "DOR2" )

}
\seealso{
\code{\link{set.custom_control}} Setting and making custom elemental classes is necessary to make a custom control class.

\code{\link{make.custom_control}} As above.

\code{\link{e_show}} For the show functions
}

