% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_data.R
\name{conditions.write_string}
\alias{conditions.write_string}
\alias{store.csv}
\alias{store.global}
\title{Store data from a simulation}
\usage{
store.csv( data, model_object, iterate, csv_name )

store.global(data, model_object, iterate, object_name)
}
\arguments{
\item{data}{A \code{simulation_analysis_object}.}

\item{model_object}{A \code{model} object.}

\item{iterate}{A numeric corresponding to which iteration of the condition list generated the data. Defaults to \code{NULL}. Either \code{iterate} or \code{csv_name}/\code{object_name} must be specified.}

\item{object_name}{For \code{store.global}. A string specifying the variable label that the data will be stored as in the global environment. Defaults to \code{NULL}. See details.}

\item{csv_name}{For \code{store.csv}. A character specifying the file name for the \code{.csv} file. Defaults to \code{NULL}. See details.}
}
\description{
The data from a simulation is typically stored in a \code{simulation_analysis_object} object, which is typically constructed by the \code{tidy.simulation_data} function. There are two methods for storing the data from an experiment: \code{store.csv} for storing the data in a \code{.csv} folder and \code{store.wd} for storing the \code{simulation_analysis_object} in the global environment.
}
\details{
When writing to a \code{.csv} file with \code{store.csv}, the \code{.csv} file will contain:

1) An \eqn{n x 2} matrix of strings with the first column listing the \code{control} objects required in the simulation and the second column containing strings that specify th parameters for the corresponding \code{control} object.

2) An \eqn{n x 2} matrix of the data in time-event format.

By default, the name of the file will be "(model_object_name)_iterate_X.csv", where X is the iterate.

When storing the data in the global environment with \code{store.global}, the  \code{simulation_analysis_object} is stored directly in the global environment with the name specified by \code{object_name} or the default "(model_object_name)_iterate_X" if \code{object_name} is unspecified.
In general, when conducting a large scale simulation, it is not advised to use \code{store.global} because you will probably not have enough RAM to store everything that you want.
}
\seealso{
\code{\link{class.analysis_object}} for \code{simulation_analysis_object}.
}

