% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/tabulation.R
\name{simple_factorial_counts}
\alias{simple_factorial_counts}
\alias{tabulation}
\title{Tabulation}
\usage{
simple_factorial_counts(event, time, rft_labels, component_label,
  response_labels, max_rfts)
}
\description{
Do this later
}
