% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.input_monitor.R
\docType{methods}
\name{trim.input}
\alias{trim.input}
\alias{trim.input,input-method}
\title{Trim the \code{input} object}
\usage{
trim.input(input_object, trim_excess_time = NULL, dims = NULL)

\S4method{trim.input}{input}(input_object, trim_excess_time = NULL,
  dims = NULL)
}
\arguments{
\item{input_object}{An object of class \code{input}.}

\item{trim_excess_time}{A numeric value. Events and responses that occur later than some specified time are also trimmed. Defaults to \code{NULL} which means that only the \code{NaN} valus are trimmed.}

\item{dims}{A character vector specifying the variables of the \code{input} object to keep. Defaults to \code{NULL} which means that all variables are kept.}

\item{input_object}{A \code{input} object.}
}
\description{
When running a simulation, it is best to set the \code{\link{input}} to contain vectors longer than what is needed to store all of the information. After running the simulation, the \code{trim.input} function can be used to trim off the unused elements.
}
\seealso{
\code{\link{class.input}}

\code{\link{tidy.simulation_data}} which uses \code{trim.input}.
}

