% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.local_rates.R
\name{compute.local_counts}
\alias{compute.local_counts}
\alias{local_rates}
\alias{compute.local_rates}
\title{Compute local rates}
\usage{
compute.local_counts(data, event_name, marker, event_offset = 0,
  marker_offset)

compute.local_rates(local_times, visit_lengths, n_bins, bin_size)
}
\arguments{
\item{data}{Either an object of class \code{simulation_analysis_object}, \code{analysis_object}, or \code{dataset}.}

\item{marker}{A variable with which to calculate the local rates within.}

\item{marker_offset}{The duration of the \code{marker_offset}. Defaults to 0.}

\item{x_event}{A variable with which to calculate the local rates over.}

\item{x_offset}{The duration of the \code{x_event}. Defaults to 0.}
}
\description{
\code{compute.local_rates} is a method for calculating the within-session local rate of some variable. Currently, methods are defined for the classes \code{analysis_object} and its child class \code{simulation_analysis_object}.
}
\details{
\subsection{\code{x_offset} and \code{marker_offset}}{
If the \code{x_event} or the \code{marker} event have non-zero durations, the \code{x_offset} and \code{marker_offset} allow for the correction of the \code{x_event}s in each bin. For example, if the local response rate is defined by the reinforcement deliveries and the reinforcement delivery lasts 3 seconds, set \code{marker_offset = -3} to subtract out the reinforcement time.
}
\subsection{Exceptions}{
In situations where there is only one \code{marker} event, \code{compute.local_rates} returns a single \code{Inf}. This would be the case if the user were to calculate the inter-reinforcement response times, but only one reinforcement delivery occurred during a particular simulation.
}

\subsection{Method for the \code{dataset} class}{
When \code{compute.local_rates} is called on a \code{dataset} class, it calls the appropriate \code{compute.local_rates} for each element in the \code{analysis_objects} slot
}
}
\examples{
#Suppose "d" is your data in an "analysis_object", "resp_time" is the response time, "rft_time" is the reinforcement time, reinforcement is delivered for 3 units of time and the operandum takes 1 unit of time to operate.
#Compute the local response rate accounting for the time to make a response and the time taken up by reinforcment
local_resp_rate = compute.local_rates( d, x_event = "resp_time", marker = "rft_time", x_event_offset = -1, marker_offset = -3 )

}
\seealso{
\code{\link{class.analysis_object}} For constructing arguments for \code{data} parameter.
}
