% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBD_utilities.R
\docType{methods}
\name{bin2int}
\alias{EBD_set}
\alias{EBD_set,EBD,integer,integer,missing-method}
\alias{bin2int}
\alias{int2bin}
\title{Evolutionary Behaviour Dynamics Utilities}
\usage{
bin2int(binaries)

int2bin(integers, digits)

EBD_set(EBD_model, phenotypes, index, genotypes)

\S4method{EBD_set}{EBD,integer,integer,missing}(EBD_model, phenotypes, index)
}
\description{
This is the documentation for the EBD utility functions.
}
\section{\code{bin2int}}{
{
    Convert a binary string from a binary vector to an integer. Works for matrices where each column is a binary string.
    \subsection{Usage}{
        \code{bin2int( binaries )}
    }
    \subsection{Arguments}{
        \describe{
            \item{\code{binaries}}{A \code{numeric} vector where each element corresponds to a binary digit, or a \code{matrix} where each column corresponds to a binary digit.}
        }
    }
    \subsection{Value}{
        Returns a \code{numeric} vector of the corresponding integers
    }
}
}

\section{\code{int2bin}}{
{
    Converts integers to their corresponding binary representation. Note that the binary string is reversed. For example \code{1} is \code{10} as opposed to \code{01}.
    \subsection{Usage}{
        \code{int2bin( integers, digits )}
    }
    \subsection{Arguments}{
        \describe{
            \item{\code{integers}}{A \code{numeric} vector of integers}
            \item{\code{digits}}{The number of digits for the binary string}
        }
    }
    \subsection{Value}{
        Returns a \code{matrix} where each column is a binary string and each row is the digit on the given string.
    }
}
}

\section{\code{EBD_set}}{
{
    Special functions for inserting behaviours into the \code{organism} slot of a \code{EBD} model. Note that this is specific to the implementation of \code{EBD} in the \code{CAB} package.
    \subsection{Usage}{
        \code{EBD_set( EBD_model, phynotypes, index, genotypes }
    }
    \subsection{Arguments}{
        \describe{
            \item{\code{EBD_model}}{A class \code{EBD} object.}
            \item{\code{phenotypes}}{A \code{numeric} vector representing behaviours.}
            \item{\code{index}}{A \code{numeric} vector of indices.}
            \item{\code{genotypes}}{A \code{matrix} with each column corresponding to a binary string of a behaviour.}
        }
    }
    \subsection{Value}{
        Modifies the \code{genotypes} and \code{phenotypes} in the \code{organism} slot of a \code{EBD} object.
    }
    \subsection{Notes}{
    Only one of \code{phenotypes} or \code{genotypes} must be specified. When one argument (e.g. \code{phenotypes} is specified, the corresponding behaviours in the \code{genotypes} are also changed accordingly.
    \code{index} is an optional argument. If it is missing, the entire population is replaced.
    The following variables must be in the \code{organism} slot for these methods to work:  "genotypes", "phenotypes", "domain", "pop_size"
    }
}
}

